requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
azure-ai-documentintelligence==1.0.2
azure-core==1.36.0
azure-cosmos==4.14.0
azure-identity==1.25.1
azure-search-documents==11.6.0
backoff==2.2.1
backports.asyncio.runner==1.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
deepeval==3.6.9
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
execnet==2.1.1
faiss-cpu==1.12.0
fastapi==0.120.1
frozenlist==1.8.0
google-auth==2.42.0
google-genai==1.46.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.3
langchain-core==1.0.1
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langfuse==3.8.1
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.38
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.2.6
ollama==0.6.0
openai==2.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
portalocker==3.2.0
posthog==6.7.11
propcache==0.4.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyfiglet==1.0.4
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.5
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pywin32==311
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tomli==2.3.0
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
